% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-methods.R
\name{confint.semmcci}
\alias{confint.semmcci}
\title{Monte Carlo Confidence Intervals for the Parameter Estimates}
\usage{
\method{confint}{semmcci}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{semmcci}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Monte Carlo Confidence Intervals for the Parameter Estimates
}
\examples{
library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- Tal.Or

# Fit Model in lavaan ------------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"

fit <- sem(data = df, model = model)

# Monte Carlo --------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)
confint(unstd)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{method}
