% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-methods.R
\name{confint.semmccistd}
\alias{confint.semmccistd}
\title{Monte Carlo Confidence Intervals for the Standardized Parameter Estimates}
\usage{
\method{confint}{semmccistd}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{semmccistd}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Monte Carlo Confidence Intervals for the Standardized Parameter Estimates
}
\examples{
library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- Tal.Or

# Fit Model in lavaan ------------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"

fit <- sem(data = df, model = model, fixed.x = FALSE)

# Monte Carlo --------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)

# Standardized Monte Carlo -------------------------------------------------
std <- MCStd(unstd, alpha = 0.05)
confint(std)

}
\seealso{
Other Monte Carlo in Structural Equation Modeling Functions: 
\code{\link{MCMI}()},
\code{\link{MCStd}()},
\code{\link{MC}()},
\code{\link{coef.semmccistd}()},
\code{\link{coef.semmcci}()},
\code{\link{confint.semmcci}()},
\code{\link{print.semmccistd}()},
\code{\link{print.semmcci}()},
\code{\link{summary.semmccistd}()},
\code{\link{summary.semmcci}()},
\code{\link{vcov.semmccistd}()},
\code{\link{vcov.semmcci}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Monte Carlo in Structural Equation Modeling Functions}
\keyword{method}
