% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-mc.R
\name{MC}
\alias{MC}
\title{Monte Carlo Confidence Intervals}
\usage{
MC(
  object,
  R = 20000L,
  alpha = c(0.001, 0.01, 0.05),
  decomposition = "eigen",
  pd = TRUE,
  tol = 1e-06,
  seed = NULL
)
}
\arguments{
\item{object}{Object of class \code{lavaan}.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.}

\item{decomposition}{Character string.
Matrix decomposition of the sampling variance-covariance matrix
for the data generation.
If \code{decomposition = "chol"}, use Cholesky decomposition.
If \code{decomposition = "eigen"}, use eigenvalue decomposition.
If \code{decomposition = "svd"}, use singular value decomposition.}

\item{pd}{Logical.
If \code{pd = TRUE},
check if the sampling variance-covariance matrix
is positive definite using \code{tol}.}

\item{tol}{Numeric.
Tolerance used for \code{pd}.}

\item{seed}{Integer.
Random seed for reproducibility.}
}
\value{
Returns an object of class \code{semmcci} which is
a list with the following elements:
\describe{
\item{\code{R}}{Number of Monte Carlo replications.}
\item{\code{alpha}}{Significance level \eqn{\alpha} specified.}
\item{\code{lavaan}}{\code{lavaan} object.}
\item{\code{decomposition}}{Matrix decomposition
used to generate multivariate
normal random variates.}
\item{\code{thetahat}}{Parameter estimates \eqn{\hat{\theta}}.}
\item{\code{thetahatstar}}{Sampling distribution of parameter estimates
\eqn{\hat{\theta}^{\ast}}.}
}
}
\description{
Calculates Monte Carlo confidence intervals
for free and defined parameters.
}
\details{
A sampling distribution of parameter estimates is generated
from the multivariate normal distribution
using the parameter estimates and the sampling variance-covariance matrix.
Confidence intervals for free and defined parameters
are generated using the simulated sampling distribution.
Parameters can be defined using the \verb{:=} operator
in the \code{lavaan} model syntax.
}
\examples{
library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- Tal.Or

# Fit Model in lavaan ------------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"

fit <- sem(data = df, model = model)

# Monte Carlo --------------------------------------------------------------
MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)

}
\references{
Pesigan, I. J. A., & Cheung, S. F. (2023).
Monte Carlo confidence intervals for the indirect effect with missing data.
\emph{Behavior Research Methods}.
\doi{10.3758/s13428-023-02114-4}
}
\seealso{
Other Monte Carlo in Structural Equation Modeling Functions: 
\code{\link{MCMI}()},
\code{\link{MCStd}()},
\code{\link{coef.semmccistd}()},
\code{\link{coef.semmcci}()},
\code{\link{confint.semmccistd}()},
\code{\link{confint.semmcci}()},
\code{\link{print.semmccistd}()},
\code{\link{print.semmcci}()},
\code{\link{summary.semmccistd}()},
\code{\link{summary.semmcci}()},
\code{\link{vcov.semmccistd}()},
\code{\link{vcov.semmcci}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Monte Carlo in Structural Equation Modeling Functions}
\keyword{mc}
\keyword{semmcci}
