% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-mc-mi.R
\name{MCMI}
\alias{MCMI}
\title{Monte Carlo Confidence Intervals (Multiple Imputation)}
\usage{
MCMI(
  object,
  R = 20000L,
  alpha = c(0.001, 0.01, 0.05),
  decomposition = "eigen",
  pd = TRUE,
  tol = 1e-06,
  adj = FALSE,
  seed_mc = NULL,
  seed_mi = NA,
  fun = "mice",
  imp = NULL,
  ...
)
}
\arguments{
\item{object}{Object of class \code{lavaan}.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.}

\item{decomposition}{Character string.
Matrix decomposition of the sampling variance-covariance matrix
for the data generation.
If \code{decomposition = "chol"}, use Cholesky decomposition.
If \code{decomposition = "eigen"}, use eigenvalue decomposition.
If \code{decomposition = "svd"}, use singular value decomposition.}

\item{pd}{Logical.
If \code{pd = TRUE},
check if the sampling variance-covariance matrix
is positive definite using \code{tol}.}

\item{tol}{Numeric.
Tolerance used for \code{pd}.}

\item{adj}{Logical.
If \code{adj = TRUE},
use Li, Raghunathan, and Rubin (1991)
sampling covariance matrix adjustment.
If \code{adj = FALSE},
use the multivariate version of Rubin's (1987)
sampling covariance matrix.}

\item{seed_mc}{Integer.
Random seed for the Monte Carlo method.}

\item{seed_mi}{Integer.
Random seed for multiple imputation.}

\item{fun}{Character string.
Multiple imputation function.
If \code{fun = "mice"}, use \code{\link[mice:mice]{mice::mice()}}.
If \code{fun = "amelia"}, use \code{\link[Amelia:amelia]{Amelia::amelia()}}.}

\item{imp}{Optional argument.
A list of multiply imputed data sets.}

\item{...}{Additional arguments to pass to \code{fun}.
If \code{fun = "mice"}, DO NOT supply \code{data}, \code{seed}, or \code{print}.
If \code{fun = "amelia"}, DO NOT supply \code{x} or \code{p2s}.}
}
\value{
Returns an object of class \code{semmcci} which is
a list with the following elements:
\describe{
\item{\code{R}}{Number of Monte Carlo replications.}
\item{\code{alpha}}{Significance level \eqn{\alpha} specified.}
\item{\code{lavaan}}{\code{lavaan} object.}
\item{\code{decomposition}}{Matrix decomposition
used to generate multivariate
normal random variates.}
\item{\code{thetahat}}{Parameter estimates \eqn{\hat{\theta}}.}
\item{\code{thetahatstar}}{Sampling distribution of parameter estimates
\eqn{\hat{\theta}^{\ast}}.}
}
}
\description{
Calculates Monte Carlo confidence intervals
for free and defined parameters.
Missing values are handled using multilple imputation.
}
\details{
A sampling distribution of parameter estimates is generated
from the multivariate normal distribution
using the parameter estimates and the sampling variance-covariance matrix
obtained using multiple imputation.
Confidence intervals for free and defined parameters
are generated using the simulated sampling distribution.
Parameters can be defined using the \verb{:=} operator
in the \code{lavaan} model syntax.
}
\examples{
library(semmcci)
library(lavaan)

# Data with Missing Values -------------------------------------------------
data("Tal.Or", package = "psych")
df <- mice::ampute(Tal.Or)$amp

# Fit Model in lavaan ------------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"

# Fit the model in lavaan using the default listwise deletion method.
fit <- sem(data = df, model = model)

# Monte Carlo --------------------------------------------------------------
MCMI(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05,
  m = 5 # use a large value e.g., 100L for actual research
)

}
\references{
Pesigan, I. J. A., & Cheung, S. F. (2023).
Monte Carlo confidence intervals for the indirect effect with missing data.
\emph{Behavior Research Methods}.
\doi{10.3758/s13428-023-02114-4}
}
\seealso{
Other Monte Carlo in Structural Equation Modeling Functions: 
\code{\link{MCStd}()},
\code{\link{MC}()},
\code{\link{coef.semmccistd}()},
\code{\link{coef.semmcci}()},
\code{\link{confint.semmccistd}()},
\code{\link{confint.semmcci}()},
\code{\link{print.semmccistd}()},
\code{\link{print.semmcci}()},
\code{\link{summary.semmccistd}()},
\code{\link{summary.semmcci}()},
\code{\link{vcov.semmccistd}()},
\code{\link{vcov.semmcci}()}
}
\concept{Monte Carlo in Structural Equation Modeling Functions}
\keyword{mc}
\keyword{semmcci}
