% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-print.R
\name{print.semmcci_std}
\alias{print.semmcci_std}
\title{Print Method for Object of Class \code{semmcci_std}}
\usage{
\method{print}{semmcci_std}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{semmcci_std}.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments.}
}
\value{
Returns a matrix of estimates, standard errors and confidence intervals.
}
\description{
Print Method for Object of Class \code{semmcci_std}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{method}
