% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.std}
\alias{print.std}
\title{Print Value}
\usage{
\method{print}{std}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Value
}
