% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-package.R
\docType{package}
\name{semmcci-package}
\alias{semmcci-package}
\alias{_PACKAGE}
\title{semmcci: Monte Carlo Confidence Intervals in Structural Equation Modeling}
\description{
Monte Carlo confidence intervals for free and defined parameters in models fitted in the structural equation modeling package 'lavaan' can be generated using the 'semmcci' package. 'semmcci' has two main functions, namely, MC() and MCStd(). The output of 'lavaan' is passed as the first argument to the MC() function to generate Monte Carlo confidence intervals. Monte Carlo confidence intervals for the standardized estimates can also be generated by passing the output of the MC() function to the MCStd() function. Preacher and Selig (2012) \doi{10.1080/19312458.2012.679848}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/semmcci}
  \item \url{https://jeksterslab.github.io/semmcci/}
  \item Report bugs at \url{https://github.com/jeksterslab/semmcci/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID}) [contributor]
}

}
\keyword{internal}
