% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci.R
\docType{package}
\name{semmcci}
\alias{semmcci}
\title{semmcci: Monte Carlo Confidence Intervals}
\description{
\code{semmcci} calculates Monte Carlo confidence intervals
for parameters defined using the \verb{:=} operator
in the structural equation modeling package \code{lavaan}.
The output of \code{lavaan} is passed as the first argument
to the \code{mc} function in \code{semmcci}
to generate Monte Carlo confidence intervals
for defined parameter/s.
}
\author{
Ivan Jacob Agaloos Pesigan
}
