% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-methods.R
\name{summary.semmcci}
\alias{summary.semmcci}
\title{Summary Method for an Object of Class \code{semmcci}}
\usage{
\method{summary}{semmcci}(object, digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{semmcci}.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of estimates, standard errors,
number of Monte Carlo replications, and confidence intervals.
}
\description{
Summary Method for an Object of Class \code{semmcci}
}
\examples{
library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- Tal.Or

# Fit Model in lavaan ------------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"

fit <- sem(data = df, model = model)

# Monte Carlo --------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)
summary(unstd)

}
\seealso{
Other Monte Carlo in Structural Equation Modeling Functions: 
\code{\link{MCMI}()},
\code{\link{MCStd}()},
\code{\link{MC}()},
\code{\link{coef.semmccistd}()},
\code{\link{coef.semmcci}()},
\code{\link{confint.semmccistd}()},
\code{\link{confint.semmcci}()},
\code{\link{print.semmccistd}()},
\code{\link{print.semmcci}()},
\code{\link{summary.semmccistd}()},
\code{\link{vcov.semmccistd}()},
\code{\link{vcov.semmcci}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Monte Carlo in Structural Equation Modeling Functions}
\keyword{method}
