% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-methods.R
\name{vcov.semmccistd}
\alias{vcov.semmccistd}
\title{Sampling Covariance Matrix of the Standardized Parameter Estimates}
\usage{
\method{vcov}{semmccistd}(object, ...)
}
\arguments{
\item{object}{Object of class \code{semmccistd}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the variance-covariance matrix
of standardized parameter estimates.
}
\description{
Sampling Covariance Matrix of the Standardized Parameter Estimates
}
\examples{
library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- Tal.Or

# Fit Model in lavaan ------------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"

fit <- sem(data = df, model = model, fixed.x = FALSE)

# Monte Carlo --------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)

# Standardized Monte Carlo -------------------------------------------------
std <- MCStd(unstd, alpha = 0.05)
vcov(std)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{method}
