% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-methods.R
\name{coef.semmcci}
\alias{coef.semmcci}
\title{Parameter Estimates}
\usage{
\method{coef}{semmcci}(object, ...)
}
\arguments{
\item{object}{Object of class \code{semmcci}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of parameter estimates.
}
\description{
Parameter Estimates
}
\examples{
library(semmcci)
library(lavaan)

# Data ---------------------------------------------------------------------
data("Tal.Or", package = "psych")
df <- Tal.Or

# Fit Model in lavaan ------------------------------------------------------
model <- "
  reaction ~ cp * cond + b * pmi
  pmi ~ a * cond
  indirect := a * b
  direct := cp
  total := cp + (a * b)
"

fit <- sem(data = df, model = model)

# Monte Carlo --------------------------------------------------------------
unstd <- MC(
  fit,
  R = 100L, # use a large value e.g., 20000L for actual research
  alpha = 0.05
)
coef(unstd)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{method}
