% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{Monte Carlo Confidence Intervals}
\usage{
mc(object, R = 20000L, alpha = c(0.001, 0.01, 0.05), plot = TRUE)
}
\arguments{
\item{object}{lavaan object
with defined parameters using the \verb{:=} operator
in the \link{lavaan} model syntax.}

\item{R}{Integer.
Number of Monte Carlo replications.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.
Default value is \code{alpha = c(0.001, 0.01, 0.05)}.}

\item{plot}{Logical.
If \code{TRUE}, plots the sampling distribution of the defined parameter estimate/s.}
}
\value{
Returns a matrix with the following columns:
\describe{
\item{est}{Estimate of indirect effect.}
\item{se}{Standard error of Monte Carlo simulated indirect effect.}
\item{limits}{Confidence limits \eqn{\frac{\alpha}{2} , 1 - \frac{\alpha}{2}}.}
}
}
\description{
Calculates Monte Carlo confidence intervals
for parameters defined using the \verb{:=} operator in \link{lavaan}.
}
\examples{
library(semmcci)
library(lavaan)

n <- 1000
x <- rnorm(n = n)
m <- 0.50 * x + rnorm(n = n)
y <- 0.25 * x + 0.50 * m + rnorm(n = n)
data <- data.frame(
  x,
  m,
  y
)
model <- "
  y ~ cp * x + b * m
  m ~ a * x
  ab := a * b
"
object <- sem(
  data = data,
  model = model
)

# Monte Carlo

mc(
  object = object,
  R = 20000L,
  alpha = c(0.001, 0.01, 0.05),
  plot = TRUE
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Monte Carlo method functions}
\keyword{mc}
