% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-print.R
\name{print.semmcci}
\alias{print.semmcci}
\title{Print Method for Object of Class \code{semmcci}}
\usage{
\method{print}{semmcci}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{semmcci}.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Method for Object of Class \code{semmcci}
}
\author{
Ivan Jacob Agaloos Pesigan
}
