% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semmcci-package.R
\docType{package}
\name{semmcci-package}
\alias{semmcci-package}
\alias{_PACKAGE}
\title{semmcci: Monte Carlo Confidence Intervals in Structural Equation Modeling}
\description{
Monte Carlo confidence intervals for free and defined parameters in models fitted in the structural equation modeling package 'lavaan' can be generated using the 'semmcci' package. 'semmcci' has three main functions, namely, MC(), MCMI(), and MCStd(). The output of 'lavaan' is passed as the first argument to the MC() function or the MCMI() function to generate Monte Carlo confidence intervals. Monte Carlo confidence intervals for the standardized estimates can also be generated by passing the output of the MC() function or the MCMI() function to the MCStd() function. A description of the package and code examples are presented in Pesigan and Cheung (2023) \doi{10.3758/s13428-023-02114-4}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/semmcci}
  \item \url{https://jeksterslab.github.io/semmcci/}
  \item Report bugs at \url{https://github.com/jeksterslab/semmcci/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID}) [contributor]
}

}
\keyword{internal}
